uriNormRules.json: tune the resolve properties and fix the periodo rule
Previously the resolve properties created URLs aimed at being as close
as possible to the normalized URI but still allowing to fetch metadata
in RDF. Unfortunatelly resolving such URLs may still cause a few
redirects which increases the time needed to fetch the RDF metadata. Now
the redirect properties are tuned to generate a direct RDF metadata
fetch URL requiring no redirects. The impact as messured on running the
uri-normalizer tests suite is cutting the retrieval time in half so it
looks like it's worht it.

Also for the wikidata retrieval the "?flavor=dump" is used now - see
https://www.wikidata.org/wiki/Wikidata:Data_access#Less_verbose_RDF_output